\name{mlt-package}
\docType{package}
\alias{mlt-package}
\title{General Information on the \pkg{mlt} Package}
\description{
  The \pkg{mlt} package implements maximum likelihood estimation in 
  conditional transformation models as introduced by Hothorn et al. (2018).

  An introduction to the package is available in the \code{mlt} package
  vignette from package \code{mlt.docreg} (Hothorn, 2018).

  A short talk on most likely transformations is available from
  \url{https://channel9.msdn.com/Events/useR-international-R-User-conference/useR2016/Most-Likely-Transformations}.

}
\references{

  Torsten Hothorn, Lisa Moest, Peter Buehlmann (2018), Most Likely
  Transformations, \emph{Scandinavian Journal of Statistics}, \bold{45}(1),
  110--134, \doi{10.1111/sjos.12291}.

  Torsten Hothorn (2018), Most Likely Transformations: The mlt Package,
  \emph{Journal of Statistical Software}, forthcoming. URL: 
  \url{https://cran.r-project.org/package=mlt.docreg}.

}
\author{
  This package is authored by Torsten Hothorn <Torsten.Hothorn@R-project.org>.
}
\keyword{package}
